Exported final product as package because WebGL didn't totally work. Please look at this, if you don't mind! :)
Had a lot of weird errors in WebGL that don't show up for me when I use the play button in Unity. A lot of these may be due to aspect ratio issues. Looks pretty great in Unity, though. :)

P.S. Feel free to adjust the hearts/lives (in GameManagerObject) or MinSpeed (in Player/PlayerMovementScript) in the Inspector if you want to modulate the difficulty for yourself. :)

Screenshots & more in a few hours! And thank you. For everything. ^~^